Rename PolynomialSampler to GeometricSampler
